backend/requirements.txt
annotated-types
anyio
certifi
chardet
click
dnspython
email_validator
fastapi
fastapi-cli
h11
httpcore
httptools
httpx
idna
Jinja2
jsonpatch
jsonpointer
langchain-core
langgraph
langgraph-checkpoint
langgraph-sdk
langsmith
markdown-it-py
MarkupSafe
mdurl
numpy
orjson
ormsgpack
packaging
pydantic
pydantic_core
Pygments
python-dateutil
python-dotenv
python-multipart
PyYAML
requests

requests-toolbelt
rich
rich-toolkit
scipy
setuptools
shellingham
six
sniffio
starlette
tenacity
typer
typing-inspection
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets

# azure api stuff
asyncio
openai
pyaudio
pygame
azure-cognitiveservices-speech


# Azure OpenAI and LLM dependencies
openai==1.91.0
python-dotenv==1.1.1
